+++ plwordnetMobileService/src/main/java/com/pwr/bzapps/plwordnetmobile/service/database/entity/grammar/PartOfSpeechEntity.java
@@ -11,25 +11,25 @@ import javax.persistence.*;
 @Table(name = "part_of_speech")
 public class PartOfSpeechEntity {
     @Id
-    private Integer id;
+    private Long id;
     @Column(name = "name_id")
-    private Integer name_id;
+    private Long name_id;
     @Column(name = "color")
     private String color;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getName_id() {
+    public Long getName_id() {
         return name_id;
     }
 
-    public void setName_id(Integer name_id) {
+    public void setName_id(Long name_id) {
         this.name_id = name_id;
     }
