+++ plwordnetMobileService/src/main/java/com/pwr/bzapps/plwordnetmobile/service/database/repository/application/ApplicationLocalisedStringRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ApplicationLocalisedStringRepository extends CrudRepository<ApplicationLocalisedStringEntity, Integer> {
+public interface ApplicationLocalisedStringRepository extends CrudRepository<ApplicationLocalisedStringEntity, Long> {
 
     @Query("SELECT als FROM ApplicationLocalisedStringEntity als WHERE id = :id AND language = :language")
-    public ApplicationLocalisedStringEntity findByIdAndLang(@Param("id") Integer id, @Param("language") String language);
+    public ApplicationLocalisedStringEntity findByIdAndLang(@Param("id") Long id, @Param("language") String language);
 
     @Query(value = "SELECT CONCAT(als.id,',','\"',' ','\"',',','\"',als.language,'\"') FROM application_localised_string als", nativeQuery = true)
     public List<String> findAllAndParseString();
@@ -21,12 +21,12 @@ public interface ApplicationLocalisedStringRepository extends CrudRepository<App
 
     @Query(value = "SELECT CONCAT(als.id,',','\"',' ','\"',',','\"',als.language,'\"') FROM application_localised_string als" +
             " WHERE als.id>=:begin AND als.id<:end", nativeQuery = true)
-    public List<String> findAllAndParseStringBatch(@Param("begin") Integer begin, @Param("end") Integer end);
+    public List<String> findAllAndParseStringBatch(@Param("begin") Long begin, @Param("end") Long end);
 
     @Query(value = "SELECT CONCAT(als.id,',','\"',' ','\"',',','\"',als.language,'\"') FROM application_localised_string als WHERE language = :language" +
             " AND als.id>=:begin AND als.id<:end", nativeQuery = true)
-    public List<String> findAllByLanguageAndParseStringBatch(@Param("language") String language, @Param("begin") Integer begin, @Param("end") Integer end);
+    public List<String> findAllByLanguageAndParseStringBatch(@Param("language") String language, @Param("begin") Long begin, @Param("end") Long end);
 
     @Query(value = "SELECT MAX(id) FROM application_localised_string", nativeQuery = true)
-    public Integer getMaxIndex();
+    public Long getMaxIndex();
 }
