+++ plwordnetMobileService/src/main/java/com/pwr/bzapps/plwordnetmobile/service/database/repository/relation/RelationTypeAllowedLexiconRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface RelationTypeAllowedLexiconRepository extends CrudRepository<RelationTypeAllowedLexiconEntity, Integer> {
+public interface RelationTypeAllowedLexiconRepository extends CrudRepository<RelationTypeAllowedLexiconEntity, Long> {
     @Query(value = "SELECT CONCAT(rtal.relation_type_id,',',rtal.lexicon_id) FROM relation_type_allowed_lexicons rtal", nativeQuery = true)
     public List<String> findAllAndParseString();
 }
